+++ Auction_shared/src/main/java/com/auction/shared/factory/ElectronicsFactory.java
@@ -1,12 +1,14 @@
 package com.auction.shared.factory;
 
-import com.auction.shared.models.Item;
+import java.math.BigDecimal;
+
 import com.auction.shared.models.Electronics;
+import com.auction.shared.models.Item;
 
 public class ElectronicsFactory implements ItemFactory{
 
     @Override
-    public Item createItem(String name, String description, double startingPrice, double currentPrice) {
+    public Item createItem(String name, String description, BigDecimal startingPrice, BigDecimal currentPrice) {
         return new Electronics(name, description, startingPrice, currentPrice);
     }
 }
