+++ Auction_shared/src/main/java/com/auction/shared/models/Vehicle.java
@@ -1,11 +1,13 @@
 package com.auction.shared.models;
 
+import java.math.BigDecimal;
+
 public class Vehicle extends Item{
     private String licensePlate;
     private int mileage;
     private int manufacturingYear;
 
-    public Vehicle(String name, String description, double startingPrice, double currentPrice){
+    public Vehicle(String name, String description, BigDecimal startingPrice, BigDecimal currentPrice){
         super(name, description, startingPrice, currentPrice);
     }
